requirements.txt
numpy>=1.23.0
scipy>=1.10.0
pandas>=1.5.0
opencv-python>=4.7.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
xformers>=0.0.20
pytorch-lightning>=2.0.0
kornia>=0.6.10
einops>=0.6.0
matplotlib>=3.5.0
PyYAML>=6.0
yacs>=0.1.8
omegaconf>=2.3.0
tqdm>=4.65.0
joblib>=1.2.0
loguru>=0.6.0
h5py>=3.0.0
requests